+++ neo/framework/File.h
@@ -71,7 +71,7 @@ public:
 	// Causes any buffered data to be written to the file.
 	virtual void			Flush();
 	// Seek on a file.
-	virtual int				Seek( long offset, fsOrigin_t origin );
+	virtual int				Seek( int offset, fsOrigin_t origin );
 	// Go back to the beginning of the file.
 	virtual void			Rewind();
 	// Like fprintf.
@@ -177,7 +177,7 @@ public:
 	virtual int				Tell() const;
 	virtual void			ForceFlush();
 	virtual void			Flush();
-	virtual int				Seek( long offset, fsOrigin_t origin );
+	virtual int				Seek( int offset, fsOrigin_t origin );
 	
 	// Set the given length and don't allow the file to grow.
 	void					SetMaxLength( size_t len );
@@ -259,7 +259,7 @@ public:
 	virtual int				Tell() const;
 	virtual void			ForceFlush();
 	virtual void			Flush();
-	virtual int				Seek( long offset, fsOrigin_t origin );
+	virtual int				Seek( int offset, fsOrigin_t origin );
 	
 private:
 	idStr					name;			// name of the file
@@ -291,7 +291,7 @@ public:
 	virtual int				Tell() const;
 	virtual void			ForceFlush();
 	virtual void			Flush();
-	virtual int				Seek( long offset, fsOrigin_t origin );
+	virtual int				Seek( int offset, fsOrigin_t origin );
 	
 	// returns file pointer
 	idFileHandle			GetFilePtr()
@@ -320,7 +320,7 @@ public:
 	virtual int				Read( void* buffer, int len );
 	
 	virtual int				Tell() const;
-	virtual int				Seek( long offset, fsOrigin_t origin );
+	virtual int				Seek( int offset, fsOrigin_t origin );
 	
 private:
 	uint64				internalFilePos;
@@ -353,7 +353,7 @@ public:
 	virtual int				Tell() const;
 	virtual void			ForceFlush();
 	virtual void			Flush();
-	virtual int				Seek( long offset, fsOrigin_t origin );
+	virtual int				Seek( int offset, fsOrigin_t origin );
 	
 private:
 	idStr					name;			// name of the file in the pak
@@ -397,7 +397,7 @@ public:
 		return 0;
 	}
 	virtual int				Tell() const;
-	virtual int				Seek( long offset, fsOrigin_t origin );
+	virtual int				Seek( int offset, fsOrigin_t origin );
 	void					SetResourceBuffer( byte* buf )
 	{
 		resourceBuffer = buf;
